+++ packages/intl-listformat/rollup.config.js
@@ -1,10 +1,19 @@
 import {uglify} from 'rollup-plugin-uglify';
 import resolve from 'rollup-plugin-node-resolve';
 import testRollupConfig from '../../rollup.config';
+import commonjs from 'rollup-plugin-commonjs';
+import json from '@rollup/plugin-json';
+const jsonConfig = json()
+const commonjsConfig = commonjs({
+  namedExports: {
+    lodash: ['pickBy', 'isEmpty', 'isEqual', 'fromPairs'],
+  },
+});
 const resolveConfig = resolve({
   mainFields: ['module', 'main'],
 });
 const uglifyConfig = uglify();
+
 export default [
   {
     input: './lib/index.js',
@@ -15,7 +24,7 @@ export default [
       exports: 'named',
       name: 'IntlListFormat',
     },
-    plugins: [resolveConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig],
   },
   {
     input: './lib/index.js',
@@ -26,7 +35,7 @@ export default [
       exports: 'named',
       name: 'IntlListFormat',
     },
-    plugins: [resolveConfig, uglifyConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig, uglifyConfig],
   },
   {
     input: './lib/locales.js',
@@ -37,7 +46,7 @@ export default [
       exports: 'named',
       name: 'IntlListFormat',
     },
-    plugins: [resolveConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig],
   },
   {
     input: './lib/locales.js',
@@ -48,7 +57,7 @@ export default [
       exports: 'named',
       name: 'IntlListFormat',
     },
-    plugins: [resolveConfig, uglifyConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig, uglifyConfig],
   },
   {
     input: './lib/polyfill-locales.js',
@@ -57,7 +66,7 @@ export default [
       file: 'dist/polyfill-with-locales.js',
       format: 'iife',
     },
-    plugins: [resolveConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig],
   },
   {
     input: './lib/polyfill.js',
@@ -66,7 +75,7 @@ export default [
       file: 'dist/umd/polyfill-intl-listformat.js',
       format: 'umd',
     },
-    plugins: [resolveConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig],
   },
   {
     input: './dist-es6/polyfill-locales.js',
@@ -77,7 +86,7 @@ export default [
       exports: 'named',
       name: 'IntlListFormat',
     },
-    plugins: [resolveConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig],
   },
   ...testRollupConfig,
 ];