+++ packages/intl-locale/rollup.config.js
@@ -2,12 +2,18 @@ import {uglify} from 'rollup-plugin-uglify';
 import resolve from 'rollup-plugin-node-resolve';
 import json from '@rollup/plugin-json';
 import testRollupConfig from '../../rollup.config';
+import commonjs from 'rollup-plugin-commonjs';
 
 const resolveConfig = resolve({
   mainFields: ['module', 'main'],
 });
 const uglifyConfig = uglify();
 const jsonConfig = json();
+const commonjsConfig = commonjs({
+  namedExports: {
+    lodash: ['pickBy', 'isEmpty', 'isEqual', 'fromPairs'],
+  },
+});
 export default [
   {
     input: './lib/index.js',
@@ -18,7 +24,7 @@ export default [
       exports: 'named',
       name: 'IntlLocale',
     },
-    plugins: [resolveConfig, jsonConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig],
   },
   {
     input: './lib/index.js',
@@ -29,7 +35,7 @@ export default [
       exports: 'named',
       name: 'IntlLocale',
     },
-    plugins: [resolveConfig, jsonConfig, uglifyConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig, uglifyConfig],
   },
   {
     input: './lib/polyfill.js',
@@ -38,7 +44,7 @@ export default [
       file: 'dist/umd/polyfill.js',
       format: 'umd',
     },
-    plugins: [resolveConfig, jsonConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig],
   },
   {
     input: './dist-es6/polyfill.js',
@@ -47,7 +53,7 @@ export default [
       file: 'dist-es6/umd/polyfill.js',
       format: 'umd',
     },
-    plugins: [resolveConfig, jsonConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig],
   },
   ...testRollupConfig,
 ];