+++ packages/intl-pluralrules/rollup.config.js
@@ -1,6 +1,14 @@
 import {uglify} from 'rollup-plugin-uglify';
 import resolve from 'rollup-plugin-node-resolve';
 import testRollupConfig from '../../rollup.config';
+import commonjs from 'rollup-plugin-commonjs';
+import json from '@rollup/plugin-json';
+const jsonConfig = json()
+const commonjsConfig = commonjs({
+  namedExports: {
+    lodash: ['pickBy', 'isEmpty', 'isEqual', 'fromPairs'],
+  },
+});
 const resolveConfig = resolve({
   mainFields: ['module', 'main'],
 });
@@ -15,7 +23,7 @@ export default [
       exports: 'named',
       name: 'IntlPluralRules',
     },
-    plugins: [resolveConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig],
   },
   {
     input: './lib/index.js',
@@ -26,7 +34,7 @@ export default [
       exports: 'named',
       name: 'IntlPluralRules',
     },
-    plugins: [resolveConfig, uglifyConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig, uglifyConfig],
   },
   {
     input: './lib/polyfill.js',
@@ -35,7 +43,7 @@ export default [
       file: 'dist/umd/polyfill.js',
       format: 'umd',
     },
-    plugins: [resolveConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig],
   },
   {
     input: './dist-es6/polyfill-locales.js',
@@ -44,7 +52,7 @@ export default [
       file: 'dist/umd/polyfill-with-locales.js',
       format: 'umd',
     },
-    plugins: [resolveConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig],
   },
   {
     input: './dist-es6/polyfill-locales.js',
@@ -55,7 +63,7 @@ export default [
       exports: 'named',
       name: 'IntlPluralRules',
     },
-    plugins: [resolveConfig],
+    plugins: [resolveConfig, commonjsConfig, jsonConfig],
   },
   ...testRollupConfig,
 ];