+++ packages/intl-relativetimeformat/scripts/extract-relative.ts
@@ -7,18 +7,10 @@
 
 import * as DateFields from 'cldr-dates-full/main/en/dateFields.json';
 import * as NumberFields from 'cldr-numbers-full/main/en/numbers.json';
-import {Locale} from './types';
-import generateFieldExtractorFn from './utils';
 import {sync as globSync} from 'glob';
 import {resolve, dirname} from 'path';
 import {FieldData, LocaleFieldsData} from '@formatjs/intl-utils';
-
-const dateFieldsLocales = globSync('*/dateFields.json', {
-  cwd: resolve(
-    dirname(require.resolve('cldr-dates-full/package.json')),
-    './main'
-  ),
-}).map(dirname);
+import * as AVAILABLE_LOCALES from 'cldr-core/availableLocales.json';
 
 // The set of CLDR date field names that are used in FormatJS.
 const FIELD_NAMES = [
@@ -59,7 +51,7 @@ export function getAllLocales() {
   }).map(dirname);
 }
 
-function loadRelativeFields(locale: Locale): LocaleFieldsData {
+function loadRelativeFields(locale: string): LocaleFieldsData {
   const fields = (require(`cldr-dates-full/main/${locale}/dateFields.json`) as typeof DateFields)
     .main[locale as 'en'].dates.fields;
   let nu: string | null = null;
@@ -120,14 +112,11 @@ function transformFieldData(data: Fields['week']): FieldData {
   return processed;
 }
 
-function hasDateFields(locale: Locale): boolean {
-  return dateFieldsLocales.includes(locale);
+export function extractRelativeFields(
+  locales: string[] = AVAILABLE_LOCALES.availableLocales.full
+): Record<string, LocaleFieldsData> {
+  return locales.reduce((all: Record<string, LocaleFieldsData>, locale) => {
+    all[locale] = loadRelativeFields(locale);
+    return all;
+  }, {});
 }
-
-const extractRelativeFields = generateFieldExtractorFn<LocaleFieldsData>(
-  loadRelativeFields,
-  hasDateFields,
-  getAllLocales()
-);
-
-export default extractRelativeFields;