+++ packages/react-intl/src/formatters/message.ts
@@ -18,7 +18,7 @@ import IntlMessageFormat, {
   FormatXMLElementFn,
   PrimitiveType,
 } from 'intl-messageformat';
-import {MessageFormatError, MissingTranslationError} from '../error';
+import {MissingTranslationError, MessageFormatError} from '../error';
 
 function setTimeZoneInOptions(
   opts: Record<string, Intl.DateTimeFormatOptions>,
@@ -100,7 +100,7 @@ export function formatMessage(
   >,
   state: Formatters,
   messageDescriptor?: MessageDescriptor,
-  values?: Record<string, PrimitiveType>
+  values?: Record<string, PrimitiveType | FormatXMLElementFn<string, string>>
 ): string;
 export function formatMessage(
   {
@@ -125,14 +125,12 @@ export function formatMessage(
   >,
   state: Formatters,
   messageDescriptor: MessageDescriptor = {id: ''},
-  values:
-    | Record<
-        string,
-        | React.ReactNode
-        | Date
-        | FormatXMLElementFn<React.ReactNode, React.ReactNode>
-      >
-    | undefined
+  values?: Record<
+    string,
+    | PrimitiveType
+    | React.ReactNode
+    | FormatXMLElementFn<React.ReactNode, React.ReactNode>
+  >
 ): React.ReactNode {
   const {id, defaultMessage} = messageDescriptor;