+++ packages/react-intl/src/index.ts
@@ -68,4 +68,4 @@ export {default as FormattedRelativeTime} from './components/relative';
 export {default as FormattedPlural} from './components/plural';
 export {default as FormattedMessage} from './components/message';
 export {createIntlCache} from './utils';
-export {ReactIntlError, ReactIntlErrorCode} from './error';
+export * from './error';