+++ packages/eslint-plugin-formatjs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.8](https://github.com/formatjs/formatjs/compare/eslint-plugin-formatjs@2.3.7...eslint-plugin-formatjs@2.3.8) (2020-05-28)
+
+**Note:** Version bump only for package eslint-plugin-formatjs
+
+
+
+
+
+## [2.3.7](https://github.com/formatjs/formatjs/compare/eslint-plugin-formatjs@2.3.6...eslint-plugin-formatjs@2.3.7) (2020-05-27)
+
+**Note:** Version bump only for package eslint-plugin-formatjs
+
+
+
+
+
 ## [2.3.6](https://github.com/formatjs/formatjs/compare/eslint-plugin-formatjs@2.3.5...eslint-plugin-formatjs@2.3.6) (2020-05-25)
 
 **Note:** Version bump only for package eslint-plugin-formatjs