+++ packages/intl-relativetimeformat/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.1](https://github.com/formatjs/formatjs/compare/@formatjs/intl-relativetimeformat@5.2.0...@formatjs/intl-relativetimeformat@5.2.1) (2020-05-28)
+
+**Note:** Version bump only for package @formatjs/intl-relativetimeformat
+
+
+
+
+
+# [5.2.0](https://github.com/formatjs/formatjs/compare/@formatjs/intl-relativetimeformat@5.1.0...@formatjs/intl-relativetimeformat@5.2.0) (2020-05-27)
+
+
+### Features
+
+* **formatjs-extract-cldr-data:** rm this package ([62bdd32](https://github.com/formatjs/formatjs/commit/62bdd32aadef899228a5303e01865f69fd729fa3))
+
+
+
+
+
 # [5.1.0](https://github.com/formatjs/formatjs/compare/@formatjs/intl-relativetimeformat@5.0.1...@formatjs/intl-relativetimeformat@5.1.0) (2020-05-25)